WIP: client connection is simulated serverside
The UnixStream closing on the client will close it on the server.
I have neglected to write the opening of a new connection as
a mirror of the client sides behavior because ssh-agent doesn't require
such behavior. The server opens a new connection right after it closes
the first one. This is because ssh-agent doesn't allow duplicate
queries on the same connection.